requirements.txt
asgiref==3.7.2
Django==3.2.21
django-cors-headers==4.2.0
djangorestframework==3.13.1
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
